vehicle select dropdown now working on sale add form
