+++ src/components/modal/vehicleDropdown.css
@@ -1,0 +1,56 @@
+/* ############ 
+    CUSTOM DROPDOWN STARTS HERE 
+############ */
+.vehicle--label {
+    margin-top: 20px;
+}
+
+.vehicle--search {
+    border: 1px solid white;
+    border-bottom: 1px solid lightgray;
+    width: 421px;
+    height: 35px;
+    line-height: 27px;
+}
+
+.vehicle--search:focus {
+    outline: none;
+}
+
+.vehicles--dropdown {
+    max-width: 421px;
+    overflow: scroll;
+    -ms-overflow-style: none;
+    /* IE and Edge */
+    scrollbar-width: none;
+    /* Firefox */
+    border: 2px solid gray;
+    border-radius: 4px;
+    height: 35px;
+    transition: height 0.4s ease;
+}
+
+.vehicles--dropdown::-webkit-scrollbar {
+    display: none;
+}
+
+.vehicles--dropdown.open {
+    height: 220px;
+    outline: none;
+    border-color: #6a78d1;
+    box-shadow: 0 0 8px #6a78d1;
+    background-color: #fff;
+}
+
+.vehicles-results--container {
+    padding: 10px;
+}
+
+.vehicles--select:hover {
+    cursor: pointer;
+    color: #6a78d1;
+}
+
+/* ############ 
+    CUSTOM DROPDOWN ENDS HERE 
+############ */
