+++ src/components/modal/vehicleSearch.js
@@ -1,42 +1,53 @@
 import React, { useState, useEffect } from "react";
 import DataManager from "../../api/dataManager";
-import VehicleDropdown from "./vehicleDropdown"
+import "./vehicleDropdown.css";
 
-const VehicleSearch = (props) => {
+const VehicleDropdown = (props) => {
 
     const [vehicles, setVehicles] = useState([]);
     const [showVehicles, setShowVehicles] = useState(false);
     const [query, setQuery] = useState("");
-    //
+    
     const handleCloseVehicleSearch = () => {
-        console.log("hello");
         setShowVehicles(false);
     };
-    //
+    
     const handleVehicleSearch = (evt) => {
         if (evt.target.value.length > 0) {
             setQuery(evt.target.value);
             setShowVehicles(true);
-            DataManager.getAll("vehicles", "vehicle", evt.target.value).then(
-                (matchedVehicles) => {
-                    console.log(matchedVehicles)
+            DataManager.getAll("vehicles", "vehicle", evt.target.value)
+                .then(matchedVehicles => {
                     setVehicles(matchedVehicles);
                 }
             );
         } else {
             setQuery("");
             setVehicles([]);
+            setShowVehicles(false);
         }
     };
 
-    // if (props.postedSale !== undefined) {
-    //     setQuery("");
-    // }
-    //
+    const handleVehicleSelect = (evt) => {
+        setQuery("");
+
+        const stateToChange = props.state
+        stateToChange.vehicle_id = parseInt(evt.target.id);
+        stateToChange.price = parseFloat(evt.target.title);
+
+        console.log(evt.target.innerHTML.split("<")[0]);
+
+        props.setSelectedVehicle(
+            {
+                price: parseFloat(evt.target.title),
+                vehicleName: evt.target.innerHTML.split("<")[0]
+            })
+
+        document.querySelector(".vehicles--dropdown").scrollTop = 0;
+    };
 
     useEffect(() => {
         props.setSelectedVehicle("");
-        // setQuery("");
 
         document.querySelector(".vehicle--search").value = "";
 
@@ -44,28 +55,49 @@ const VehicleSearch = (props) => {
 
     return (
         <>
-            <label style={{marginTop: "20px"}} className="name--label">Select Vehicle:</label>
-            <input
-                className="modal--input vehicle--search"
-                type="text"
-                onChange={handleVehicleSearch}
-                value={props.selectedVehicle.vehicleName ? props.selectedVehicle.vehicleName : query }
-                placeholder="Search Vehicles"
-            />
-            {showVehicles ? (
-                <VehicleDropdown
-                vehicles={vehicles}
-                handleCloseVehicleSearch={handleCloseVehicleSearch}
-                showVehicles={showVehicles}
-                setShowVehicles={setShowVehicles}
-                state={props.newSale}
-                setSelectedVehicle={props.setSelectedVehicle}
-                setState={props.setNewSale}
-                setQuery={setQuery}
-                {...props} />
-                ): null}
+            <label className="vehicle--label">Select Vehicle:</label>
+            
+            <div 
+                className={`vehicles--dropdown ${showVehicles ? "open" : ""}`}
+                onBlur={handleCloseVehicleSearch}
+            >
+                <input
+                    className="vehicle--search"
+                    type="text"
+                    onChange={handleVehicleSearch}
+                    value={props.selectedVehicle.vehicleName ? props.selectedVehicle.vehicleName : query }
+                    placeholder="Search Vehicles"
+                />
+
+                {vehicles.length > 0 ? (
+                    <div className="vehicles-results--container">
+                        {vehicles.map((vehicle) => {
+                            return (
+                                <>
+                                    <div
+                                        className="vehicles--select"
+                                        id={vehicle.id}
+                                        title={vehicle.floor_price}
+                                        onClick={handleVehicleSelect}
+                                    >
+                                        {`${vehicle.make} ${vehicle.model}`}
+                                        <span
+                                            className="vin"
+                                            id={vehicle.id}
+                                            title={vehicle.floor_price}
+                                            style={{pointerEvents: "none"}}
+                                        >
+                                            #{vehicle.vin}
+                                        </span>
+                                    </div>
+                                </>
+                            );
+                        })}
+                    </div>
+                ) : null}
+            </div>
 
         </>
     )
 }
-export default VehicleSearch;
+export default VehicleDropdown;
