+++ src/components/sale/list.css
@@ -76,44 +76,3 @@
     padding: 0 !important;
     margin-bottom: 15px;
 }
-
-
-.vehicles--dropdown.open {
-    margin-top: 20px;
-    background-color: #F5F8FA;
-    box-shadow: 1px 3px 11px 7px #a5a0a0;
-    padding: 20px;
-    border-radius: 8px;
-    overflow: scroll;
-    z-index: 100000;
-    width: 600px;
-    max-height: 400px;
-    display: flex;
-    flex-flow: wrap;
-}
-
-.vehicles--select {
-    display: flex;
-    flex-direction: column;
-    width: 159px;
-    padding: 5px;
-    margin-right: 27px;
-    margin-bottom: 30px;
-    overflow: scroll;
-    border: 1px solid black;
-    border-radius: 7px;
-    background-color: #fff;
-    height: 100px;
-}
-.vehicles--select:nth-of-type(3n) {
-    margin-right: 0;
-}
-.vehicles--select:hover {
-    cursor: pointer;
-    border: 1px solid #FF7A59;
-}
-
-.vin {
-    font-size: 11px;
-    color: gray;
-}
